+++ Decorator.cc
@@ -1,4 +1,4 @@
-#include "decorator.h"
+#include "Decorator.h"
 #include <memory>
 
 using namespace std;
