+++ GameStateManager.cc
@@ -1,4 +1,4 @@
-#include "gamestatemanager.h"
+#include "GameStateManager.h"
 
 GameStateManager::GameStateManager(){}
 GameStateManager::GameStateManager(std::shared_ptr<GameStateManager> gsm): stockManager{gsm->stockManager}, userManager{gsm->userManager} {}
