+++ GamerunUI.cc
@@ -1,4 +1,4 @@
-#include "gamerun.h"
+#include "GamerunUI.h"
 #include "ui_gamerun.h"
 #include <iostream>
 #include <vector>
