+++ main.cc
@@ -1,4 +1,4 @@
-#include "mainwindow.h"
+#include "MainWindow.h"
 #include <QApplication>
 
 int main(int argc, char *argv[])
